+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### Fixes and Changes
 
+- Module [8FACE mk2](./docs/EightFaceMk2.md)
+    - Added "Bind module (select multiple)" option (#291)
 - Module [GLUE](./docs/Glue.md)
     - Changed "Add label" hotkey to Ctrl+G (#305)
 - Module [MIDI-KEY](./docs/MidiKey.md)